Document specification and requirements using OpenSpec (#24)
This PR uses [OpenSpec](https://openspec.dev/) to document
specifications and requirements. The documents are intended to be read
by humans to understand the design choices and details of this project.
They also serve as a standardized way to prompt AI agents if needed.

This PR also updates the contribution rules to include mandatory
disclosure of use of AI. Note that, before this PR (or version 0.0.3),
only certain tests are generated by AI, which had been reviewed by the
owner in each PR.

In this PR, the `project.md` is written by hand, and the specs are
generated by AI from existing codebase.